:recycle: [refactory]: Update Directory
- Update Directory (login of pages.)
