+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { useRecoilValue } from 'recoil';
-import MainLoading from './components/MainLoading';
 
-import { managerAuth } from './pages/login/context/login';
-import { useMe } from './pages/login/custom_hooks/useMe';
+import { managerAuth } from './pages/login/application/context/login';
+import { useMe } from './pages/login/application/custom_hooks';
+import MainLoading from './components/MainLoading';
 import LoginRouter from './router/LoginRouter';
 import MainRouter from './router/MainRouter';
